+++ mayan/apps/common/warnings.py
@@ -7,6 +7,12 @@ class DatabaseWarning(UserWarning):
     """
 
 
+class DeprecationWarning(UserWarning):
+    """
+    Warning when a feature or interface has been deprecated
+    """
+
+
 class InterfaceWarning(UserWarning):
     """
     Warning when using obsolete internal interfaces